pyproject.toml
[tool.poetry]
name = "lingorank-interface"
version = "0.1.0"
description = "Dash interface for presenting difficulty estimation and simplification results for French texts"
authors = ["Henri Jamet <henri.jamet@unil.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
pandas = "^2.2.1"
dash = "^2.16.1"
dash-mantine-components = "^0.12.1"
transformers = "^4.38.2"
torch = "^2.2.1"
peft = "^0.9.0"
flask = "^3.0.2"
black = "^24.2.0"
python-git = "^2018.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
